update trigger-client to send correct payload
